# CompilerWidgetClient/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "CompilerWidgetApi",
    "version": "1.0"
  },
  "paths": {
    "/api/compile/project/{projectId}/run": {
      "post": {
        "tags": [
          "Compiler"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RunRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/compile/project/{projectId}/stop": {
      "post": {
        "tags": [
          "Compiler"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/compile/project/{projectId}/status": {
      "get": {
        "tags": [
          "Compiler"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStatus"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStatus"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcessStatus"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/compile/project/{projectId}/compile": {
      "post": {
        "tags": [
          "Compiler"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResult"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResult"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompileResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{fileId}/save": {
      "post": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{fileId}/rename": {
      "post": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{fileId}/delete": {
      "post": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/files/project/{projectId}": {
      "post": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RenameFileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    },
    "/api/files/{fileId}": {
      "post": {
        "tags": [
          "File"
        ],
        "parameters": [
          {
            "name": "fileId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CompilationError": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "string",
            "nullable": true
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "startLine": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "endLine": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CompileRequest": {
        "type": "object",
        "properties": {
          "mainFile": {
            "type": "string",
            "nullable": true
          },
          "optimize": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "CompileResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "output": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompilationError"
            },
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "compiledAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "ProcessStatus": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "isRunning": {
            "type": "boolean"
          },
          "processId": {
            "type": "string",
            "nullable": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "uptime": {
            "type": "string",
            "format": "date-span"
          }
        },
        "additionalProperties": false
      },
...
      "RenameFileDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "RunRequest": {
        "type": "object",
        "properties": {
          "mainFile": {
            "type": "string",
            "nullable": true
          },
          "environmentVariables": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "timeoutSeconds": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RunResult": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "output": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompilationError"
            },
            "nullable": true
          },
          "processId": {
            "type": "string",
            "nullable": true
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateFileDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}